+++ courses/management/tests/migrate_edx_data_test.py
@@ -1,4 +1,5 @@
-"""Tests for migrate_edx_data management command's repair_migrated_profiles type"""
+"""Tests for migrate_edx_data management command's repair_migrated_profiles
+type and --dry-run behavior"""
 
 import pytest
 
@@ -183,3 +184,61 @@ def test_limit_caps_number_of_users_repaired():
 
     repaired = User.objects.filter(legal_address__isnull=False).count()
     assert repaired == 1
+
+
+def test_migrate_users_dry_run_creates_no_records(capsys):
+    """--dry-run must not create any User/LegalAddress/UserProfile rows"""
+    existing_user = UserFactory.create(email="existing@example.com")
+    conn = FakeConnection(
+        columns=["user_email", "user_full_name"],
+        rows=[
+            ("new1@example.com", "New One"),
+            ("new2@example.com", "New Two"),
+            (existing_user.email, "Existing User"),
+        ],
+    )
+
+    Command()._migrate_users(conn, {"dry_run": True})  # noqa: SLF001
+
+    assert User.objects.count() == 1  # only the pre-existing user
+    output = capsys.readouterr().out
+    assert "[DRY RUN] Would create 2 users" in output
+
+
+def test_migrate_users_dry_run_respects_batching(capsys):
+    """The dry-run count must accumulate correctly across multiple fetchmany
+    batches, not just within a single batch
+    """
+    conn = FakeConnection(
+        columns=["user_email", "user_full_name"],
+        rows=[(f"new{i}@example.com", f"New {i}") for i in range(5)],
+    )
+
+    Command()._migrate_users(conn, {"dry_run": True, "batch_size": 2})  # noqa: SLF001
+
+    assert User.objects.count() == 0
+    output = capsys.readouterr().out
+    assert "[DRY RUN] Would create 5 users" in output
+
+
+def test_migrate_users_real_run_creates_user_records():
+    """Without --dry-run, matching rows actually create User rows.
+
+    Deliberately not asserting on legal_address here: _bulk_create_users
+    calls User.objects.bulk_create(..., ignore_conflicts=True), and Django
+    never populates .pk on returned objects when ignore_conflicts=True is
+    used, on any backend - confirmed empirically against this test DB. That
+    means _bulk_create_legal_addresses/_bulk_create_user_profiles, which
+    filter on those (always-None) ids, never actually create anything today.
+    That's a separate, pre-existing bug unrelated to --dry-run - flagged
+    separately, not fixed here.
+    """
+    conn = FakeConnection(
+        columns=["user_email", "user_full_name"],
+        rows=[("new@example.com", "New User")],
+    )
+
+    Command()._migrate_users(conn, {})  # noqa: SLF001
+
+    user = User.objects.get(email="new@example.com")
+    assert user.name == "New User"
